+++ src/DigiSign.php
@@ -34,7 +34,7 @@ final class DigiSign implements EndpointInterface
 {
     public const VERSION = '2.9.0';
     public const API_BASE = 'https://api.digisign.org';
-    public const API_BASE_TESTING = 'https://api.digisign.digital.cz';
+    public const API_BASE_TESTING = 'https://api.testing.digisign.org';
 
     /** The base URL for requests */
     private string $apiBase = self::API_BASE;