+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `AccountBilling.accountContactLimit` property
 - Add `AccountBilling.accountContactUsage` property
 - Add `ContactImportsEndpoint` and `ContactImportItemsEndpoint`
+- Add `BrandingInfo` resource
+- Add `AccountBrandingsEndpoint.info()` method
 
 ## [2.9.0] - 2025-10-03
 - [BREAKING-CHANGE] Fix `WebhookAttempt.status` type from string to int