+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `BrandingInfo` resource
 - Add `AccountBrandingsEndpoint.info` endpoint
 - Add `User.invitedAt` property
+- Specify `BatchSending.importFields` property type
 
 ## [2.9.0] - 2025-10-03
 - [BREAKING-CHANGE] Fix `WebhookAttempt.status` type from string to int